// src/query.graphql.ts
export interface PullRequestQuery_repository_pullRequest {
  __typename: "PullRequest";
  /**
   * Identifies the pull request number.
   */
  number: number;
  /**
   * Identifies the state of the pull request.
   */
  state: PullRequestState;
  /**
   * Whether or not the pull request can be merged based on the existence of merge conflicts.
   */
  mergeable: MergeableState;
  /**
   * Detailed information about the current pull request merge state status.
   */
  mergeStateStatus: MergeStateStatus;
  /**
   * The commit that GitHub automatically generated to test if this pull request could be merged. This field will not return a value if the pull request is merged, or if the test merge commit is still being generated. See the `mergeable` field for more details on the mergeability of the pull request.
   */
  potentialMergeCommit: PullRequestQuery_repository_pullRequest_potentialMergeCommit | null;
  /**
   * A list of reviews associated with the pull request.
   */
  reviews: PullRequestQuery_repository_pullRequest_reviews | null;
  /**
   * A list of labels associated with the object.
   */
  labels: PullRequestQuery_repository_pullRequest_labels | null;
  /**
   * Identifies the pull request title.
   */
  title: string;
  /**
   * The body as Markdown.
   */
  body: string;
  /**
   * Author's association with the subject of the comment.
   */
  authorAssociation: CommentAuthorAssociation;
  /**
   * Identifies the base Ref associated with the pull request.
   */
  baseRef: PullRequestQuery_repository_pullRequest_baseRef | null;
  /**
   * Identifies the oid of the base ref associated with the pull request, even if the ref has been deleted.
   */
  baseRefOid: any;
  /**
   * A list of commits present in this pull request's head branch not present in the base branch.
   */
  commits: PullRequestQuery_repository_pullRequest_commits;
  /**
   * A list of commits present in this pull request's head branch not present in the base branch.
   */
  allCommits: PullRequestQuery_repository_pullRequest_allCommits;
  /**
   * Identifies the head Ref associated with the pull request.
   */
  headRef: PullRequestQuery_repository_pullRequest_headRef | null;
  /**
   * Identifies the oid of the head ref associated with the pull request, even if the ref has been deleted.
   */
  headRefOid: any;
  /**
   * Identifies the name of the head Ref associated with the pull request, even if the ref has been deleted.
   */
  headRefName: string;
  /**
   * The repository associated with this node.
   */
  repository: PullRequestQuery_repository_pullRequest_repository;
}

export interface PullRequestQuery_repository {
  __typename: "Repository";
  /**
   * Returns a single pull request from the current repository by number.
   */
  pullRequest: PullRequestQuery_repository_pullRequest | null;
}

export interface PullRequestQuery {
  /**
   * Lookup a given repository by the owner and repository name.
   */
  repository: PullRequestQuery_repository | null;
}

export interface PullRequestQueryVariables {
  owner: string;
  repo: string;
  pullRequestNumber: number;
}